belvoapp/requirements.txt
django==3.2.5
djangorestframework==3.12.4
factory-boy==3.2.0
drf-yasg==1.20.0